Refactor test utils
We will soon move every test related files intout a tests folder outside
of inclusion_connect
